pyproject.toml
[tool.poetry]
name = "promptsForChat"
version = "0.1.0"
description = ""
authors = ["vale95ntino"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^1.5.2"
snscrape = "^0.4.3.20220106"
requests = "^2.28.1"
opencv-python = "^4.6.0.66"
numpy = "^1.23.5"
pillow = "^9.3.0"
pytesseract = "^0.3.10"
autocorrect = "^2.6.1"
langdetect = "^1.0.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
